Add three new electron ID DNN working points and the outputs of the DNN to derivations
Adding three new electron ID working points based on a DNN to derivations. Additionally,
the single output nodes of the DNN are saved. The additional outputs are saved in the tool
using the TResult. The outputs are only saved for the Loose wp since all wps use the same
network and have the same outputs.