tauRecTools: fix severe bug with neutral PFOs in tau substructure reconstruction
Hello,

This MR is fixing a very problematic behaviour with neutral PFOs in the tau substructure reconstruction.
When we process a tau, we check if there are EM clusters within 0.4 of the tau and if so, we create new neutral PFOs that we push to the neutral PFO container (1 neutral PFO per EM cluster).
However, when we run the substructure reconstruction for the current tau, we loop over all the PFOs from the neutral PFO container, which means we also pick up the PFOs from previous taus in the event, not only those of the current tau.
And... we overwrite the properties of all PFOs (which are non-const)...

Fortunately, for the tau itself, it is mostly OK, because the PFOs of the current tau are ~properly subtracted.
So the tau variables (including those depending on the substructure reconstruction output) are mostly OK at reconstruction time.
However, the PFOs we store in xAODs are nonsense, as all the PFOs in an event are last altered by the last tau of the event.
Also, there can be very weird situations (fortunately not too frequent) where duplicate PFOs are created, in case an EM cluster is shared by 2 taus, and in this case, the tau reconstruction itself is affected.

Adding the 'urgent' flag, as we'd really need to have this merged in the release that will be used for Round2 MC preproduction expected imminently.

Cheers,
Bertrand