introduce PythonConfig and EL::PythonConfigBase
The goal here is to introduce a single python configurable in
AnalysisBase, instead of the current scheme in which there is a
separate configurable for AnaAlgorithm, AsgTool, etc. and essentially
duplicating a lot of code.

For now this duplicates even more code.  Once this is all tested and
validated, the old code will eitehr be completely removed or made to
inherit from the new classes.