Deweight NSW hits in the TrackTruthSelector
Because there are many more hits in the NSW than there are in the other stations, they need to be deweighted here so they don't dominate the truth matching.
The weight of 0.25 is chosen as there would be 24 hits expected in the NSW and 6 in one of the other stations.
This should help to resolve the issues with poor efficiency reported in ATLASRECTS-5610.