TriggerJobOpts: refactor EDMVersion flag
Move the code to discover the `EDMVersion` flag in bytestream files to
`TrigEDMConfig.Utils` and cache the result for a given bytestream file.
This ensures we are only reading the file once. This is especially
important for mixed old/new job options where the flags are not being
locked and hence this function is called every time code reads the
`EDMVersion` flag.