Change to JetUncertainties treatment of negative mass
JetCalibTools and JetUncertainties currently have different methods of
treating the combined mass of jets where both of the input masses are
negative.  This is an extreme case, but it should be properly handled in
order to differentiate between input problems vs different conventions.

This MR changes JetUncertainties to now accept events where the two
input mass definitions (calo and TA) are both negative, and the combined
mass value is 0.  In contrast, if the calo and TA masses are both
negative while the combined mass value is non-zero, then this is a
different problem that is unrelated to JetCalibTools and most likely
signifies a problem in the user input file.