Replaced the temporary ReadDecorHandle objects with class-level Accessor-s.
As described in ATLASRECTS-6317, in a RAWtoALL reconstruction job the usage of
the ReadDecorHandle variables was showing up as a bottleneck in threading tests.

Beside replacing all ReadDecorHandle objects with Accessor ones in the code,
I also removed all __attribute__((unused)) declarations from it. Replacing them
with a much more platform independent way of not triggering warnings about
unused variables in the first place.