Fix histogram binning and remobe BeamSpotSvc use
This fixes some of the binning for histograms - the d0 and z0 error estimates
are positive definite, yet the histograms plotted a symmetric range about 0.

Also, all the use of the BeamSpotSvc has been removed - it is currently present
but protected by an #ifdef - this will be removed in the future.

In principle, since we use the xAOD::TrackParticles these are already beamline
corrected, and we calculate the beamline ourselves for the truth particles.

These changes will break the use of track collections such as Trk::Track which are
not beamline corrected, but we can add that functionality back later if required