MuidCaloEnergyTools: Fix clang warning.
Fix possible use of uninitialized variable.

This was introduced by a recent cleanup:

commit 5874eb062f7c026e6c7a1b9b1ee96e9ef819b3c1
Author: Johannes Josef Junggeburth <johannes.josef.junggeburth@CERN.CH>
Date:   Sun May 16 18:07:31 2021 +0200

    Start tidy up of CaloEnergy meausrement

The change here restores the behavior of this fragment to what it was before
this change.  However, the code still doesn't really make a lot of sense:
FinalEnergyError always ends up the same as FinalEnergyErrorPlus
(or FinalEnergyErrorMinus).  If this is what was intended, the redundant
code should be removed.