Don't access elements in the vector by adding an int to an iterator
A non-reproducible FPE is observed at this line, presumably because it is accessing garbage rather than an element of the vector.
Switching to vector.at(nextSf) might not solve the problem but should at least make it clearer what the problem is.
RunTier0Tests.py confirms no differences from this change.