AthContainers: Remove lock from fast path in AuxTypeRegistry::findAuxID.
The locking in AuxTypeRegistry::findAuxID turns out to be a bottleneck
for many-core jobs.  However, in the case where the auxid is already
defined, we don't really need to take the lock.

Add a fast path before we take the lock to look up the auxid.
If everything's ok, then we can return without taking the lock.
But if something goes wrong, then we acquire the lock and do the
full procedure.

See ATLASRECTS-6317.