AthenaServices: In MT jobs, clear the event store in an algorithm context.
In many-core jobs, clearing the store is observed to be a bottleneck,
since it is executed serially by the event loop manager.

For MT jobs, change so that the store is cleared from the event loop manager's
incident handler on an EventWritten incident.  Incidents are processed by an
algorithm, so this should allow the stores to be cleared for multiple
slots at the same time.

Also remove the ClearStoreProperty from AthenaHiveEventLoopMgr, since for
MT jobs it doesn't make sense to set this to anything other than
EndEvent.