Set flag correctly
the "first" flag was set to ensure that the range was always set for the first
time round the loop, but then this was set explicitly and the loop started from 1,
so ths meant that the firstvalues from the loop were always being overwritten by the
second values from the loop, and then only compared with the later values.

This fixes that.