TrigT1CTMonitoring: Fix retrieval of LVL1::NimCTP.
LVL1::NimCTP is now recorded as const.  This caused retrieval here to fail,
breaking q431.  Change to retrieve it as const.

Note: this code had a pre-existing const_cast used to change the object
in SG.  This is not thread-safe.  In practice, it's probably ok with
the current configuration, but this should be rethought.