TrigSerializeCnvSvc: Suppress FPEs while reading BTaggingTrigAuxContainer_v1.
Many variables in xAOD::BTaggingTrigAuxContainer_v1 were changed
from double to float.  However, we wrote data in the past which contained values
that were valid doubles but which were out of range for floats.
So we can get FPEs in TrigBSExtraction when we read them.
Disable FPEs when we're reading an instance of that class.