Fix for rare case in which duplicate hits are added to a MuGirl track
In rare cases in which there are S/L overlaps, MuGirl may reconstruct the same part of the overlap as two segments with identical hits but opposite signs on the radius.
The correct segment will be successfully combined with the other part of the overlap, but the wrong one will remain.
MuGirl will then try to reconstruct a track based on the combined segment plus the incorrect segment.
This fix catches this (extremely rare) case and prevents the track reconstruction from happening.
The combined segment can still be used to reconstruct a track: only the combination of the combined and incorrect segment is vetoed.