Major update of DAOD_PHYSVAL in master
This is a major update of DAOD_PHYSVAL in master, restoring thinning, synchronising jet/MET/b-tagging with DAOD_PHYS, dropping old r21 material. Several parts are still commented out as the relevant tools have not yet been migrated to r22.
Note to shifters: the commented lines should remain since we expect to be able to uncomment them soon and it is useful to see what is still left to be done. Since DAOD_PHYSVAL is not used for physics anlaysis yet, and since no other software uses the modified script, this will have no effect on any other output.