disable all CURRENT warnings in 21.2
As agreed in the analysis release meeting the specific goal here is
just to disable all the warnings that have existed in 21.2 for a long
time and that lead to occasional confusion during the merge reviews.
In a few cases I tried to address the actual source of the warning,
but for the most part I just picked the most expedient way to shut it
off, instead of trying to determine whether they point to an actual
problem or not, and then fixing them accordingly.