Introduced MbtsReadoutBuilder
The code which builds MBTS readout geometry has been factored out from the EndcapCryostatConstruction,
so that now it can be shared by the old-style and new-style (SQLite based) LAr detector factories.

Also fixed SqliteRecordset so that it does not print an ERROR message if the requested
table is not found in the SQLite database.