tauRec and JetRec: update LC jet seeding + tau thinning
Hello,

This MR is implementing a configuration for jet seeding + tau thinning that is closer to what we will use in the R22 reprocessing, namely:

* no JES-calibrated pt cut on the AntiKt4LCTopoJets used to seed the tau reconstruction
* a tau pt cut applied to the MVA tau pt at the end of the tau reconstruction, to thin the tau containers before writing them out to AOD

The tau thinning cut will be fine tuned, for now it's a simple 10 GeV pt cut.
This will lead to a significant increase in CPU as we will now run the tau reconstruction on all seed jets, not only on those with pt>15 GeV.
The impact on the AOD size depends on the MVA energy resolution and other factors that are still being tuned.
We expect to see changes in SPOT plots (ATLASRECTS-5684).

Cheers,
Bertrand