TrigBSExtraction: simplify configuration and code cleanup
Merge the `EFResultKey` and `HLTResultKey` properties as they serve the
same purpose (final result of EF/HLT processing).

Fix a bug in the `equalize` option. This should only be set to `true` if
reading data with an L2 and EF result (Run-1 data). In all other cases
this is not necessary (but also did not cause any harm as the L2
navigation would have been empty).

Also delete commented code and unused ResultKey properties in
`TrigBStoxAODTool`.