Update log file diff in post_tpcnvtest.sh
Similar change to !35801.

Prior to this change log file diffs in unit tests compared the reference file to the current
output, rather than the other way around, so that lines in the log removed due to local
changes are marked with a `+` in the diff output and lines added to the log are marked with
a `-`.

This commit switches the order of comparison with the aim of making the output more intuitive.