Fix FPE due to log(X) where X <= 0 in JMS calib
This FPE was first seen in r22, but is present in both r21 and r22.
This commit is the r21 fix, which can then be swept into r22.
Note that the choice was made to keep the same behaviour as the current
tool, which looked for NaNs instead of preventing their creation in the
first place.  That is, negative logarithms result in a calibration factor
of 1 (no calibration) or a calibration weight of 0 (for combined mass
calibrations).