rework AsgService interface declarations and configuration
As part of implementing my first dual-use service I decided that the
existing macros are not quite as elegant as I hoped, and also don't
match gaudi mechanisms as closely as I would have liked.

I also updated the dual-use service configuration, so that it works
nicely in a dual-use setting.