remake SysListLoaderAlg into SystematicsSvc
For now this just replaces the algorithm with a service, and
communicates the systematics list via the service interface instead of
via the event store.

The plan is to add more functionality in future merge requests that
will move a lot of the systematics handling functionality from the
configuration layer into the service.