Add J/Psi muon support triggers
Adding support triggers that requre the dimuon mass to be in a window around J/Psi mass and the dimuon pairs to have opposite charge (added support to hypo tool to apply this requirement). Also fixed a logic bug in the invariant mass hypo tool so that both the upper and lower bounds of the mass are applied to the same pair (previously, a dimuon pair could pass if it satisfied only one bound if a second pair satisified the other).
The added trigger is equivalent to the JPsimumuFS trigger we had in Run 2 (naming changed from Run 2 to make it clearer and to take into account existing naming conventions of other chains). ATR-23614.