Fix FPE due to log(X) where X <= 0 in jet mass calibration
This FPE was first seen in r22
(https://its.cern.ch/jira/browse/ATR-23551), but is present in both r21
and r22. The original commit was the r21 fix, but the automatic sweep
to r22 failed.  This is a manual sweep of the same changes, where it
looks like the sweep failed because of white-space differences between
the r21 and r22 version of the file being modified.

The original r21 MR is !44419