tauRecTools: use ghost-matching for tau-track association
Hello,

This MR is implementing a hybrid tau-track association that accepts all tracks within a cone of radius 0.25, and beyond 0.25, tracks are required to be ghost-matched with the seed jet.
The ghost matching requirement prevents from using tracks from another tau (a situation that leads to very problematic behaviour of tau substructure reconstruction, ATLTAU-1442).
The motivations are explained in [this talk](https://indico.cern.ch/event/1051117/contributions/4420815/attachments/2269038/3853037/TauCP_Martin_22Jun21.pdf)

The hybrid cone/ghost scheme is meant to replace the R=0.4 cone association, but it's currently disabled in tauRec.

I am adding the urgent flag because we need to validate the new scheme urgently and use it for upcoming retunings of tau algorithms for the R22 reprocessing.

Cheers,
Bertrand