Fix for ATLASRECTS-6436
With the more careful tests for the quality of covariance matrices, tracks are more likely to be rejected.
So this downgrades the warning for the case that tracks are rejected based on a bad covariance, while leaving it for the case that the track has no perigee or the perigee has no covariance.