tauRecTools: first R22 calibration for combined TES
Hello,

This MR is deploying a preliminary version of the combined Tau Energy Scale calibration for R22 (ATLTAU-1740), tuned on Round1+Round2 MC preproduction samples.
It uses a compatibility condition between calo-only pt and substructure pt to reduce non-Gaussian tails from substructure pt.
So far, the compatibility condition was the same for all decay modes, but it now depends on the decay mode.
The calibration is preliminary (the ROOT file is on EOS, soon on cvmfs), and a final tune will be deployed for the R22 reprocessing.

Cheers,
Bertrand