Update combined TES calibration file in client code
Hello,

This MR is a follow-up of !44709 where the structure of the calibration file for the combined tau energy scale calibration changed.
The changed was made to the tau reconstruction but not propagated to client codes.
This MR should fix the problem, until a better solution is found (client codes should try to pick the tauRec file using tauRecFlags, rather than hardcode a file name).

Cheers,
Bertrand