tauRecTools: use ghost matching by default for tau-track association
Hello,

This MR is a follow-up of !44662 that introduced ghost matching in the tau-track association procedure (ATLTAU-1442).
Ghost matching is now activated by default in the offline tau reconstruction (won't be used in the trigger).

In addition, tracks beyond R=0.25 of the tau axis and which are not ghost matched are now kept if the jet closest to the track is the seed jet of the current tau.
This condition was introduced to recover tau-track association efficiency for low-pt 3p taus, where tracks from the tau decay are not always ghost-matched with the seed jet.
This scheme will be used for the R22 reprocessing after it is carefully validated.

Sorry for adding the urgent flag again, but it would be great to have this MR reviewed and merged quickly, as it will be needed for the final Round of algorithm tuning for the R22 reprocessing.

Cheers and thanks,
Bertrand