AthenaServices: Thread-local alternate stack.
When we establish the core dump signal handler, we also try to set up
an alternate stack, so that it will work even if the stack is
corrupt/exhausted.  But this didn't really work for MT jobs because
the alternate stack is thread-local.

Make the alternate stack thread-local.  Use the existing BeginEvent
handler to ensure that the alternate stack is set for the current thread.

See ATLASRECTS-6442.