tauRec: new (temporary) default for RNN track classifier
Hello,

As discussed in ATLPHYSVAL-715, this MR is deploying a new default calibration for the RNN tau track classifier.
This is the latest available tuning that was deployed for Round2 preproduction already via a preExec.
The goal of this MR is just to avoid specifying this in future preExecs.
But it's a temporary tune, it will be superseded in a couple of weeks when we have the final tune for the R22 reprocessing.

The change was tested in tau standalone reco from ESD (and was anyway used for Round2 MC preprod).

Cheers,
Bertrand