Avoid file peeking if evgen is running
This adds a mechanism for checking if we're running evgen and avoiding
input file peeking in that case. It is not the most elegant thing I've
ever written, but it is functional. The ugliness is contained to a
single place (DerivationFrameworkMaster) which is an action-on-import
module that needs to be modified into a function call at some point to
make all of this easier - but that's a bit tricky, so here we're just
getting the job done.