tauRecTools: prepare for R22 RNN tau ID
Hello,

This MR is preparing the ground for R22 RNN tau ID, which we expect to deploy in the next days (ATLTAU-1741).
For now, we use dummy calibration files for 2p (the same files as for 3p, such that the reconstruction output is unchanged).

We also introduce new variables that will be used in the R22 RNN for tracks and clusters.
Some are close to R21 equivalents (track impact parameters: log vs no log), but the trigger still relies on R21 variable definitions, so we cannot decommission the R21 variables yet.

The next MR will just push new calibration files (new json files for the RNN, new ROOT files for the working points).

Cheers,
Bertrand