Ignore stream name of event-less files Fix ATEAM-757
The EventStreamInfo object of event-less files is, by necessity, empty.
That means dereferencing the first element of the vectory containing the
stream names that produced the file causes problems. So, ignore the
stream name when this vector is empty --- noting that this implies an
event-less files was encountered. The assumption here is that the files
was produced by the correct stream, we the stream name just cannot be
looked up here.