Select first file with events for metadata peeking
Much of the information the autoconfiguration grabs from file metadata
requires at least one event to have been written by the stream that
produced the file. To ensure those values are present in the metadata
supplied to the autoconfiguration by the peeker select use the first
input file with at least one event.