eflowRec: skip thinned taus in PFTauFEAssoc
Hello,

This MR is a follow-up of !44978 that introduced a dynamic variable to indicate whether a tau will be written to AOD.
The variable is now exploited in PFTauFlowElementAssoc where thinned taus can be skipped.
It should hopefully address both the CPU increase reported in ATLASRECTS-6412 and the invalid element links reported in ATLPHYSVAL-771.

Cheers,
Bertrand