tauRecTools: fix CombinedTES weight and deploy final calo TES calibration
Hello,

This MR adds a protection in the calculation of a weight that is meant to be within [0,1], but can be <0 or >1 in rare cases.
The weight is used in the calculation of the weighted average of calo-only tau pt and substructure tau pt.
This MR changes the tau reconstruction output in rare cases.

I also take this opportunity to deploy the final R22 calo-only calibration file tuned on Round 2.5 MC preproduction.

Cheers,
Bertrand