PanTau: fix irregular tau decoration
Hello,

This MR addresses the problem reported in ATLASRECTS-6460: a decoration is added onto some taus but not all.
When calling this earlier,
```
  tauJet->setPanTauDetail(xAOD::TauJetParameters::PanTauDetails::PanTau_DecayModeExtended, decayMode_PanTauExtended);
```
xAODCheckerSummary.py no longer complains.

The rest is just minor cleanup that does not affect the output.

Although this MR changes the tau reconstruction output (TauJetsAuxDyn.PanTau_DecayModeExtended), I'm not adding the corresponding label because it's a variable that is used nowhere, neither in the reconstruction nor in analysis.
But I do add the bugfix label :-)

Cheers,
Bertrand