Fix configuration for full scan muon trigger hypo tool
Fixing the configuration of the full scan muon trigger ypo tool to get the correct multiplicity and thresholds for multi-muon noL1 triggers (ATR-23813). This affects the counts for any triggers containing NmuXnoL1 where N is greater than one, so the chain count references have been updated accordingly.
Also improved slightly the switching on of the nscan trigger configuration (what was there worked only because we specify the nscan part of the trigger on the last chain part. Have switched to something more correct/robust).