tauRecTools: deploy final combined TES and MVA TES calibrations
Hello,

This MR is deploying the final Tau Energy Scale calibrations for combined TES (ATLTAU-1740) and MVA TES (ATLTAU-1739).
The tunes were derived on Round 2.5 gamma* MC preproduction.

For MVA TES, two small changes were made to the variable definitions.

* First, ptDetectorAxis was replaced with the proper tau-vertex-corrected LC-scale tau pt, namely ptIntermediateAxis.
These two pt calibrations are very close, but we should be using ptIntermediateAxis.
* Second, for BDT scores that come out of the tau substructure reconstruction (BDTs that separate 1p0n/1p1n/1pXn and 3p0n/3pXn modes), something had been overlooked.
The BDT scores are initialised to a default value of -1111 in the reconstruction, while the meaningful values set in the substructure reconstruction are within [-5,1].
So having a range that spans [-1111, 1] is quite bad for the MVA TES BRT optimisation, basically the cuts on these variable seemed completely random (TMVA probably scans for optimal cuts with a fixed step size).
Scanning values in the meaningful range [-5-epsilon, 1] makes much more sense, although the BRT performance does not improve much.

I'm also updating the digest counts, as tau energy calibration is known to impact the number of tracks in AODs, via tau track thinning.

Adding the urgent flag as the MR is needed for the R22 reprocessing.

Cheers,
Bertrand