Temporary hack to get the ID Trigger nightly tests working again
For some reason, the previous release is not able to correctly handle
the trigger from the current release.

In the ID Trigger, to free us from sensitivity the offline release
that ofthen fails, we set the job to use the last stable offline release
for the post processing.

However, because the old release does not work with the trigger from the
current release, these jobs fail.

It addresses, but does not ultimately fix, ATR-23775