tauRec: update ESD input file for standalone tau reconstruction
Hello,

This MR is updating the ESD input files that are used in the tau standalone reconstruction from ESD (used for ART tests).
The recently updated ESD was problematic because it was missing Kt4LCTopoOriginEventShape, which is used as input for the MVA tau energy calibration.
The event shape had been temporarily removed from ESDs but is now back: !45286 .

I am also removing the test that runs tau reconstruction seeded from EMTopo jets, as we are now certain this configuration won't be used in the next 5 years.
R22 will stick to LCTopo seed jets.

Cheers,
Bertrand