VP1HepVis: Assertion / error message fix
We were tripping the assertion in  SbPolyhedronPolygonXSect::Internals::setData
because it could be called with three points (n=3).  But that case seems
to be handled ok, so change the assertion to n>2.

We also got a bunch of errors from SbPolyhedronCons::SbPolyhedronCons
because the tracking geometry can produce zero-height cones.
Just ignore those cases.