Updating sde_strategy in MGC to old default
In versions of MadGraph newer than 2.9, there's an sde_strategy run
parameter that defines how the job does the integration. It's a change
of strategy that may behave better under some circumstances. We have
seen in the validation, AGENE-2038, that this strategy is automatically
set based on the process being used. This seems confusing, and the
automatic setting doesn't always align with what is naively expected.

After some discussion, we decided to always set the default strategy,
and of course to allow the user to change the strategy if they choose to
do so. This MR updates the run card as soon as a new process is defined
to ensure that the default strategy is used. If the user chooses to set
the new strategy they'll be able to do so, and their setting will not be
overridden. This has become AGENE-2063

I took the opportunity to clean up a couple other cases where the run
card was being modified and we were being overlay prescriptive about
what to set.

This is a manual sweep of !45406