tauRecTools: minor fix in track classifier property
This MR is simplifying the handling of a tool property. The property we are interested in is an unsigned int (maximum number of tracks passed to the tau track RNN classifier), but as far as I understand it's not a type that is explicitly supported:
https://gitlab.cern.ch/atlas/athena/-/blob/master/Control/AthToolSupport/AsgTools/AsgTools/Property.h#L26-39

That might have been the reason why a sort of workaround was implemented in the tool, reading an int, and casting it to an unsigned int afterwards.
But apparently with this workaround, even though the value ultimately used at runtime was correct, the value printed by --dump-configuration was weird (like 32765, or 32766).

The new implementation no longer shows this feature, the value is printed correctly, and ultimately the variable value is the same (correct) as before.

So I put the bugfix flag, but that's not really fixing a bug. The reconstruction output is unchanged.

Cheers,
Bertrand