Update conditions tags used in trigger MC tests
Updating to use by default OFLCOND-MC16-SDR-25-02 as the conditions tag for trigger tests running on MC (instead of determining based on the input file) as discussed in ATLASRECTS-6437.

It's still possible to override this as needed (and this is already done for tests that need different conditions (tests with NSW inputs, for example)).