RPC_CondCabling: avoid expensive copies of SectorLogicSetup
`RpcCablingCondAlg` was accidentally making copies of the huge
`SectorLogicSetup` class. Make sure to use references and move
semantics. This speeds up the creation of the conditions by a
factor ~2.5 (e.g. number `malloc`s are reduced from 11M to 3M).

To avoid future problems, make the class move-only as we don't really
need to make copies.