Adding machinery to job transforms to allow skimming of DAOD_PHYS (D2AODs) in master
This MR implements the machinery in the job transforms required to make skims of DAOD_PHYS (called D2AODs). No skims are yet implemented. This was previously implemented in release 21.2 but not in master, an oversight which is corrected here. The JT skeleton was checked out directly from release 21.2 since it isn't present in master, and the other files were updated manually to avoid over-writing more recent changes in master. Depending on the outcome of a discussion about nomenclature, this mechanism may also be used to make PHYSLITE from PHYS.