Bug fix: Use wrong vector to calculate largest bin number in findBin()
findBin returns the correct bin of the correction to use for all binned
correction types. It checks if the value lies in the first given bin,
then the second, etc. If it does not lie in any of the given bins, then
it must be in the "overflow" bin going up to "infinity". The calculation
of the overflow bin number is done incorrectly, since the wrong vector
is used: The vector saving all vectors with binned correction values,
instead of the vector saving the currently used binning.