CxxUtils: Update stacktrace test for newer lcg versions.
With newer lcg versions, the system addr2line doesn't work because it's too old.
Normally that doesn't matter, because we usually use eu-addr2line, which does
work.  However, the stacktrace unit test forces the use of /usr/bin/addr2line
(as the output formats are slightly different).

Change stacktrace so that if we don't find /usr/bin/eu-addr2line we
search PATH for addr2line, rather than hardcoding it.  Also adjust the
unit test so that it should work with either addr2line or eu-addr2line.