L1DynamicPedestalProviderTxt: migrate to BunchCrossingCondData
Migrate the pedestal corrector from `TrigBunchCrossingTool` to
`BunchCrossingCondData`. Instead of caching the bunch distance structure
just calculate it every time as this should be fast enough.

Note that the `gapBefore/AfterBunch` methods in `TrigBunchCrossingTool`
have a bug that cause it to return one extra BC. E.g. the first empty
bunch after a train would have `gapBeforeBunch()==1` whereas the
`BunchCrossingCondData` returns the correct value `0`. The
`distanceFromHeadOfTrain` function has been adapted and it was verified
that the distance values are the same before/after migration to the
conditions object.

Partially addresses ATR-23378.