tauRec: fix TauTrackFinder config when not using ghost matching
This is another submission of !45787 where the target branch is master, not 22.0-mc20.
The CI in 22.0-mc20 does not seem to be configured that well, with many tests failing, so it may be easier to review the MR within master.
If this MR is reviewed, then !45787 should be closed.

Cheers,
Bertrand