TrigHLTJetHypo+TriggerMenuMT: use exception for empty HypoTool
Throw a `NoHypoToolCreated` exception in case no HypoTool was created in
the generator function. Returning `None` as in the previous code results
in duplicate empty tool names in the ToolHandleArray (ATR-23920).

In addition remove the redundant copy of tool names and `getTools` accessors in
`MenuSequence` (the `ToolHandleArrary` already ensures we are not adding
any duplicates).

Also remove an explicitly set log-level in `Cabling.py`.