ComponentAccumulator: fix empty ToolHandle warning in conf2toConfigurable
Fix these type of warnings
```
Py:conf2toConfigurable WARNING Component: "" is of type string, no conversion, some properties possibly not set?
```
that appear when trying to convert an empty `ToolHandle` property. This
is a valid use-case and we should not be warning about that.

Also fix some of the debug message text and indentation.