Modifications to ensure RandomRunNumber does not get written in DAOD_PHYS when run in a train with PHYSLITE
It is necessary to be able to run DAOD_PHYS and PHYSLITE together in a train. The latter runs pile-up re-weighting and thus produces a transient RandomRunNumber. Due to the smart slimming option in PHYS being set to AllVariables, this led to the RandomRunNumber being written also in PHYS, which is fatal for running onward production from PHYS where pile-up re-weighting is needed.

The changes are:
- switch PHYS to use the smart slimming for EventInfo (plus some extra variables)
- remove RandomRunNumber from the EventInfo smart slimming, which was there erroneously
- Add RandomRunNumber to PHYSLITE, where it is needed for analysis-level pile-up reweighting