tauRecTools: final MVA TES tune for R22
Hello,

This is a resubmission of !46086 but targeting the 22.0-mc20 branch instead of master, to save 1 day due to the sweep.
If this MR is merged, then !46086 should be closed.
Tagging @seroslia1 who reviewed the original MR.

This MR is deploying the final MVA Tau Energy Scale calibration for the R22 Run2 reprocessing.
It was derived on Round 3 MC preproduction (ATLTAU-1785).
In the previous iteration (Round 2.5), the tau track classifier and the tau energy scale were tuned separately, while the energy scale depends on track classification.
So we needed to update MVA TES using Round 3 MC that has the final track classifier tune.
The updated MVA TES is better behaved (very slightly better resolution, and the double-Gaussian structure that was seed in the energy resolution at very high pt is now gone).
The performance validation can be found in [these slides](https://cernbox.cern.ch/index.php/s/C3csX2EjEr1yG4X).

One input variable of the 0-prong boosted regression tree was seen to be useless and was dropped (it is now only used for the >=1 prong MVA TES BRT).

The thinning cuts (pt cuts above which we write taus to AOD) were adjusted to get the same thinning efficiencies vs pt as with Round 2.5 MC, such that the AOD size is almost unchanged.
From SPOT tests, the total event size changes from 342.391 kb to 343.225 kb for ttbar MC, and from 399.221 kb to 398.751 kb for data17.

I also update the digest counts: as one could expect, we observe small changes in the number of ID tracks, due to TRT standalone tracks being thinned from AODs if they are not used by egamma nor tau reconstruction (and tau thinning is pt-dependent, thus depends on the MVA TES tune).

The MR was tested in tau standalone reco from ESD, q221, and xAOD digest.

I put the urgent flag because we would need this for the reprocessing.

Cheers,
Bertrand