Move overlap removal from muon hypo alg to hypo tool
Moving the muon overlap removal from the HypoAlg to the HypoTool, so that we only consider muons that are active for a particular chain. Updating the counts reference as this changes slightly some multi-muon chain counts where we were previously incorrectly flagging muons as overlapping with muons that were only pasing other chains. Fixes ATR-23996.