Introduced new helper function for building Calo DD manager
The code that builds the Calo DD manager has been factored out of the converter
and turned into a standalone helper function. The idea is to use this function
in both the converter and the Calo alignment condition algorithm. The latter
is yet to be developed.

This code refactoring also required the move of the CaloCellVolumes class from the
converter into the new CaloDetDescrUtils package.