Prepared Calo Descriptors and Elements to work with yet-to-be-implemented CaloAlignCondAlg
Added a pointer to GeoAlignmentStore as an optional argument to the constructors
of Calo descriptor and element objects. As a result, these objects now can access mis-aligned
positions of LAr volumes.

Also dropped several obsolete methods of the descriptor and element classes, and migrated
CaloDetDescrBuilder to the updated CaloDetDescr interface.