Ensure consistency between number of brems and brem states.
For electrons the number of brems was forced to one or zero, neglecting that
elsewhere the number of brems was determined from the states of a trajectory.
In some cases there were more than one brem-states for electron trajectories
which caused range errors. These range errors are addressed here by ensuring
that the number of brems